--- plane-api/src/main/java/com/henriquels25/airlines/PlaneApiApplication.java
@@ -1,13 +1,0 @@
-package com.henriquels25.airlines;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PlaneApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PlaneApiApplication.class, args);
-	}
-
-}
