--- plane-api/src/main/java/com/henriquels25/airlines/airport/Airport.java
@@ -1,14 +1,0 @@
-package com.henriquels25.airlines.airport;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-
-@Builder
-@AllArgsConstructor
-@Getter
-@EqualsAndHashCode
-public class Airport {
-    private String code;
-}
