--- plane-api/src/main/java/com/henriquels25/airlines/infra/controller/Error.java
@@ -1,10 +1,0 @@
-package com.henriquels25.airlines.infra.controller;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class Error {
-    private final String message;
-}
