--- plane-api/src/main/java/com/henriquels25/airlines/plane/PlaneNotFoundException.java
@@ -1,4 +1,0 @@
-package com.henriquels25.airlines.plane;
-
-public class PlaneNotFoundException extends RuntimeException {
-}
