--- plane-api/src/main/java/com/henriquels25/airlines/plane/PlaneRepository.java
@@ -1,9 +1,0 @@
-package com.henriquels25.airlines.plane;
-
-import java.util.Optional;
-
-public interface PlaneRepository {
-    String save(Plane plane);
-
-    Optional<Plane> findById(String id);
-}
