--- plane-api/src/main/java/com/henriquels25/airlines/plane/infra/controller/FlightArrival.java
@@ -1,17 +1,0 @@
-package com.henriquels25.airlines.plane.infra.controller;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.extern.jackson.Jacksonized;
-
-@Jacksonized
-@Builder
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Getter
-public class FlightArrival {
-
-    private final String planeId;
-    private final String airport;
-
-}
