--- plane-api/src/main/java/com/henriquels25/airlines/plane/infra/mongo/AirportDocument.java
@@ -1,8 +1,0 @@
-package com.henriquels25.airlines.plane.infra.mongo;
-
-import lombok.Data;
-
-@Data
-public class AirportDocument {
-    private String code;
-}
