--- plane-api/src/main/java/com/henriquels25/airlines/plane/infra/mongo/SpringPlaneRepository.java
@@ -1,6 +1,0 @@
-package com.henriquels25.airlines.plane.infra.mongo;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-interface SpringPlaneRepository extends MongoRepository<PlaneDocument, String> {
-}
