--- plane-api/src/main/java/com/henriquels25/airlines/plane/infra/stream/PlaneEventSender.java
@@ -1,7 +1,0 @@
-package com.henriquels25.airlines.plane.infra.stream;
-
-public interface PlaneEventSender {
-
-    void send(PlaneEvent planeEvent);
-
-}
