+++ plane-api/src/main/java/com/henriquels25/planeapi/PlaneApiApplication.java
@@ -1,0 +1,13 @@
+package com.henriquels25.planeapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PlaneApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PlaneApiApplication.class, args);
+	}
+
+}
