+++ plane-api/src/main/java/com/henriquels25/planeapi/airport/Airport.java
@@ -1,0 +1,14 @@
+package com.henriquels25.planeapi.airport;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+
+@Builder
+@AllArgsConstructor
+@Getter
+@EqualsAndHashCode
+public class Airport {
+    private String code;
+}
