+++ plane-api/src/main/java/com/henriquels25/planeapi/infra/controller/Error.java
@@ -1,0 +1,10 @@
+package com.henriquels25.planeapi.infra.controller;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class Error {
+    private final String message;
+}
