+++ plane-api/src/main/java/com/henriquels25/planeapi/plane/PlaneNotFoundException.java
@@ -1,0 +1,4 @@
+package com.henriquels25.planeapi.plane;
+
+public class PlaneNotFoundException extends RuntimeException {
+}
