+++ plane-api/src/main/java/com/henriquels25/planeapi/plane/PlaneOperations.java
@@ -1,0 +1,13 @@
+package com.henriquels25.planeapi.plane;
+
+import com.henriquels25.planeapi.airport.Airport;
+
+import java.util.Optional;
+
+public interface PlaneOperations {
+    String create(Plane plane);
+
+    Optional<Plane> findById(String id);
+
+    void arrivedIn(String planeId, Airport airport);
+}
