+++ plane-api/src/main/java/com/henriquels25/planeapi/plane/PlaneRepository.java
@@ -1,0 +1,9 @@
+package com.henriquels25.planeapi.plane;
+
+import java.util.Optional;
+
+public interface PlaneRepository {
+    String save(Plane plane);
+
+    Optional<Plane> findById(String id);
+}
