+++ plane-api/src/main/java/com/henriquels25/planeapi/plane/infra/controller/FlightArrival.java
@@ -1,0 +1,17 @@
+package com.henriquels25.planeapi.plane.infra.controller;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.extern.jackson.Jacksonized;
+
+@Jacksonized
+@Builder
+@JsonIgnoreProperties(ignoreUnknown = true)
+@Getter
+public class FlightArrival {
+
+    private final String planeId;
+    private final String airport;
+
+}
