+++ plane-api/src/main/java/com/henriquels25/planeapi/plane/infra/controller/ResourceNotFoundException.java
@@ -1,0 +1,8 @@
+package com.henriquels25.planeapi.plane.infra.controller;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class ResourceNotFoundException extends RuntimeException {
+}
