+++ plane-api/src/main/java/com/henriquels25/planeapi/plane/infra/mongo/AirportDocument.java
@@ -1,0 +1,8 @@
+package com.henriquels25.planeapi.plane.infra.mongo;
+
+import lombok.Data;
+
+@Data
+public class AirportDocument {
+    private String code;
+}
