+++ plane-api/src/main/java/com/henriquels25/planeapi/plane/infra/mongo/PlaneDocument.java
@@ -1,0 +1,18 @@
+package com.henriquels25.planeapi.plane.infra.mongo;
+
+import lombok.Data;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+@Document("planes")
+@Data
+class PlaneDocument {
+
+    @Id
+    private String id;
+
+    private String code;
+    private String type;
+
+    private AirportDocument airport;
+}
