+++ plane-api/src/main/java/com/henriquels25/planeapi/plane/infra/mongo/SpringPlaneRepository.java
@@ -1,0 +1,6 @@
+package com.henriquels25.planeapi.plane.infra.mongo;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+interface SpringPlaneRepository extends MongoRepository<PlaneDocument, String> {
+}
