+++ plane-api/src/main/java/com/henriquels25/planeapi/plane/infra/stream/PlaneEvent.java
@@ -1,0 +1,13 @@
+package com.henriquels25.planeapi.plane.infra.stream;
+
+import lombok.*;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Getter
+@EqualsAndHashCode
+public class PlaneEvent {
+    private String planeId;
+    private String currentAirport;
+}
