+++ plane-api/src/main/java/com/henriquels25/planeapi/plane/infra/stream/PlaneEventSender.java
@@ -1,0 +1,7 @@
+package com.henriquels25.planeapi.plane.infra.stream;
+
+public interface PlaneEventSender {
+
+    void send(PlaneEvent planeEvent);
+
+}
